pyproject.toml
[project]
name = "agenticchatbot"
version = "0.1.0"
description = "Agentic AI chatbot using LangGraph, Groq models, and Streamlit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"faiss-cpu",
	"langchain",
	"langchain-community",
	"langchain-core",
	"langchain-groq",
	"langchain-openai",
	"langchain-tavily",
	"langgraph",
	"python-dotenv",
	"streamlit",
	"tavily-python",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
